Organise Konfa::AutoDoc code into one namespace:
 * As former Konfa::AutoDoc class only did parsing, move that to designated class
 * Fold the Generator namespace into Konfa::AutoDoc
 * Make "Formatter" base class for formatters inside of the Konfa::AutoDoc namespace
 * Extract shared formatter examples to rspec module so it can be used by specs for external formatters
 * Load all things necessary by requiring "konfa/autodoc"
